refactor: 열거형으로 ButtonStyle 관리 (#11)
