+++ BBANGZIP-iOS/BBANGZIP-iOS/Sources/BbangButtonType.swift
@@ -1,0 +1,14 @@
+//
+//  BbangButtonType.swift
+//  BBANGZIP
+//
+//  Created by 김송희 on 5/15/25.
+//
+
+import SwiftUI
+
+enum BbangButtonType {
+    case primary
+    case secondary
+    case disabled
+}
